requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.3
MarkupSafe==2.1.5
pkgutil_resolve_name==1.3.10
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.29
SQLAlchemy-serializer==1.4.12
SQLAlchemy-Utils==0.41.2
typing_extensions==4.11.0
Werkzeug==3.0.2
zipp==3.18.1
